microservices/ai_service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
pydantic
pydantic-settings
prometheus-fastapi-instrumentator
langchain
langchain-google-genai
google-generativeai
fastmcp
langchain-classic
